Voxel-Build 116: Argument names and autostart
The argument system has been slightly changed to introduce
argument names. Instead of just passing a bunch of integers and
booleans now the argument value is preceded by the name, such
as -width 1280 -height 720 instead of 1280 720.

This made possible to introduce a new flag, which is -autostart.
If provided, will start the game automatically like if the Play
button was pressed instead of jumping to the menu. Good for
testing.