Build-197: Inventory and Fixes
-Changes:

Add Full Dynamic inventory, to enter press "E" and to move items use the "space" key to push, move with the arrows and pop with "lcontrol" key.
Now there is a try catch in the main loop to handle any exception.
Minor changes to physics.
Moved "menu" package to "ui" package.
Inventory are saved in the world data.
Sun now don't scale on diferents resolutions.

-BugFixes:

Fixed #59 .